requirements.txt
anthropic>=0.40.0
notion-client>=2.0.0
click>=8.1.0
python-dotenv>=1.0.0
